+++ dot_man/__init__.py
@@ -1,4 +1,4 @@
 """dot-man: Dotfile manager with git-powered branching."""
 
-__version__ = "1.2.2"
+__version__ = "1.2.3"
 __author__ = "Bishoy Ehab"